Add a --shutdown mode (#8)
When the --shutdown flag is provided, the specified command is forked
(instead of simply exec'd) and the linkerd-await program forwards
SIGTERM to the child process. When the child process completes, the
proxy is shutdown via its admin interface. This behavior only operates
as expected with as-of-yet unreleased proxy versions that support the
/shutdown admin endpoint (linkerd/linkerd2-proxy#811). On prior versions,
The /shutdown call will not succeed and the proxy will remain running.

This change also introduces a 5s ready request timeout so that readiness
requests cannot stall indefinitely.