Support the LINKERD_AWAIT_DISABLED env
The LINKERD_DISABLED environment variable can be used to disable
`linkerd-await`; but this can be misleading, as sometimes all you really
want to do is disable `linkerd-await` but not actually disable the
Linkerd proxy.

This change updates `linkerd-await` to prefer a new environment
variable, `LINKERD_AWAIT_DISABLED`, in addition to `LINKERD_DISABLED`.