Make CMD optional (#23)
`CMD` is not required if the use-case of `linkerd-await` is only to wait for readiness
but do nothing afterwards.

In linkerd/linkerd2#5967 this is the case. `linkerd-await` is executed as a container
hook and the only thing that it needs to do is prevent the hook from finishing until
the proxy is ready. Once it is ready, it exits without running any additional commands.

Signed-off-by: Kevin Leimkuhler <kevin@kleimkuhler.com>