Introduce a --verbose flag
Issue #41 reports that it's undesirable to emit an error message when
linkerd-await is disabled, as this is noisey for users who may not need
or want to know anything about linkerd-await.

This change introduces a `--verbose` flag (and an environment variable,
`LINKERD_AWAIT_VERBOSE`) that controls whether this message is emitted.
The default behavior is changed to be silent.

Fixes #41